Create User Creation Script(userscripy.sh)
A bash script to automate the creation of multiple user accounts on a CentOS server. The script reads usernames from a text file and creates corresponding user accounts with default passwords. It checks if a user already exists to avoid duplication and sets the initial password using the chpasswd command.